+++ src/Futhark/CodeGen/Backends/COpenCL.hs
@@ -44,8 +44,8 @@ sizeHeuristicsCode (SizeHeuristic platform_name device_type which (TPrimExp what
 
     which' = case which of
       LockstepWidth -> [C.cexp|ctx->lockstep_width|]
-      NumGroups -> [C.cexp|ctx->cfg->default_num_groups|]
-      GroupSize -> [C.cexp|ctx->cfg->default_group_size|]
+      NumBlocks -> [C.cexp|ctx->cfg->default_num_groups|]
+      BlockSize -> [C.cexp|ctx->cfg->default_group_size|]
       TileSize -> [C.cexp|ctx->cfg->default_tile_size|]
       RegTileSize -> [C.cexp|ctx->cfg->default_reg_tile_size|]
       Threshold -> [C.cexp|ctx->cfg->default_threshold|]
@@ -105,11 +105,6 @@ mkBoilerplate opencl_program macros kernels types failures = do
   GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_program(struct futhark_context_config *cfg, const char* s);|]
   GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_dump_binary_to(struct futhark_context_config *cfg, const char* s);|]
   GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_load_binary_from(struct futhark_context_config *cfg, const char* s);|]
-  GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_default_group_size(struct futhark_context_config *cfg, int size);|]
-  GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_default_num_groups(struct futhark_context_config *cfg, int size);|]
-  GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_default_tile_size(struct futhark_context_config *cfg, int size);|]
-  GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_default_reg_tile_size(struct futhark_context_config *cfg, int size);|]
-  GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_default_threshold(struct futhark_context_config *cfg, int size);|]
   GC.headerDecl GC.InitDecl [C.cedecl|void futhark_context_config_set_command_queue(struct futhark_context_config *cfg, typename cl_command_queue);|]
   GC.headerDecl GC.MiscDecl [C.cedecl|typename cl_command_queue futhark_context_get_command_queue(struct futhark_context* ctx);|]